Added invisible border to entry containers
Previously selecting an image would add a blue border, but this would reduce the space inside the container by 4px on all sides, making the image 8px smaller in both directions. A border is now always there but is invisible when the image is not selected, creating a gap between the container and the image and preventing image resizing when selecting an image.